+++ src/server/game/Battlegrounds/Zones/BattlegroundBFG.cpp
@@ -410,7 +410,7 @@ void BattlegroundBFG::EventPlayerClickedOnFlag(Player* source, GameObject* /*tar
     TeamId teamIndex = GetTeamIndexByTeamId(source->GetTeam());
 
     // Check if player really could use this banner, not cheated
-    if (!(m_Nodes[node] == 0 || teamIndex == m_Nodes[node] % 2))
+    if (!(m_Nodes[node] == 0 || teamIndex == uint8(m_Nodes[node] % 2)))
         return;
 
     source->RemoveAurasWithInterruptFlags(SpellAuraInterruptFlags::PvPActive);
